+++ app/helpers/editions_helper.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
 module EditionsHelper
-  def display_rating(score)
-    if score.zero?
+  def display_rating(score, walked_out: false)
+    if walked_out
+      "🚪🚶"
+    elsif score.zero?
       "💣"
     elsif score == 5.0
       "🔥"