+++ app/models/category.rb
@@ -14,6 +14,6 @@ class Category < ApplicationRecord
   # We could memoize this, but I don't think it will be heavily used
   # I'm also wary to add more side-effects to saving a Rating
   def overall_average_rating
-    ratings.average(:score).to_f
+    ratings.counting_towards_aggregates.average(:score).to_f
   end
 end