+++ db/migrate/20260517161000_add_walked_out_to_ratings.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddWalkedOutToRatings < ActiveRecord::Migration[8.0]
+  def change
+    add_column :ratings, :walked_out, :boolean, default: false, null: false
+  end
+end